requirements.txt
llama-cpp-python
whisper
pyttsx3
speechrecognition